requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.1.15
django-allauth==0.44.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-compressor==2.4.1
django-crispy-forms==1.11.2
django-filebrowser==3.12.1
django-grappelli==2.13.4
django-import-export==2.5.0
django-js-asset==1.2.3
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
rcssmin==1.0.6
requests==2.27.1
requests-oauthlib==1.3.1
rjsmin==1.1.0
six==1.16.0
tablib==3.2.0
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
